fix: renamed to containers dir
