+++ src/navigation/index.ts
@@ -1,5 +1,5 @@
-import App from '../container/App';
 import {Navigation} from 'react-native-navigation';
+import App from 'containers/App';
 
 Navigation.registerComponent('MainScreen', () => App);
 Navigation.events().registerAppLaunchedListener(() => {
